backend/requirements.txt
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database and ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Cache and queue
redis==5.2.0

# Validation and settings
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# OpenAI integration
openai==1.58.1

# Document processing
PyMuPDF==1.24.13
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2

# Text processing and chunking
tiktoken==0.8.0
nltk==3.9.1

# Vector search and embeddings
faiss-cpu
numpy==1.26.4

# Background task processing
celery==5.4.0
kombu==5.4.2

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
black==24.10.0
ruff==0.8.4
